==> week7/0-Outline.ipynb <==
## Topic Model (from webscraping to machine learning)

**Courtesy of Alice Zhao**

### Data Collection & Cleaning

- Steps
    1. **Getting the data - **in this case, we'll be scraping data from a website
    2. **Cleaning the data - **we will walk through popular text pre-processing techniques
    3. **Organizing the data - **we will organize the cleaned data into a way that is easy to input into other algorithms

- Output
    1. **Corpus** - a collection of text
    2. **Document-Term Matrix** - word counts in matrix format
        - (1) Clean Text - remove excess, unnecessary parts of the text
        - (2) Tokenize Text - split the text into smaller pieces
        - (3) Document-Term Matrix - Put into a matrix so a machine can read it


### Exploratory Data Analysis

1. **Most common words** - find these and create word clouds
2. **Size of vocabulary** - look number of unique words and also how quickly someone speaks
3. **Amount of profanity** - most common terms


### Sentiment Analysis

1. Polarity: How positive or negative a word is. -1 is very negative. +1 is very positive.
2. Subjectivity: How subjective, or opinionated a word is. 0 is fact. +1 is very much an opinion.


### Topic Modeling

1. Covering the basic steps on how to do **Latent Dirichlet Allocation (LDA)**, which is one of many topic modeling techniques.
